fix issue w/ None being cached when value is missing
Unfortunately the previous, beautiful implementation using get_or_set
does not work quite as elegantly as I'd hoped because while Django does
not store the default value when it's None, it *does* store it when it's
a callable that *returns* None.

There is a [pull request][1] to fix this behavior in Django, but since
the behavior is already out there (at least in 1.11.5) I basically can't
use get_or_set the way I want for the foreseeable future (i.e. until
it's reasonable to have Django >= 1.11.6 as a requirement).

[1]: https://github.com/django/django/pull/9087