remove FallthroughCache.create method
This method actually only existed for the tests (I wanted a friendlier
way to instantiate a FallthroughCache instance), but there's no great
reason to make it part of the public interface for the class. Any
consumer of this library would not be expected to instantiate the class
directly, but have it load automatically by Django.

Since the code is useful for the tests, I moved it there.